requirements.txt
python-dotenv
openai
gpt_index
pypdf
PyCryptodome
pandas
openpyxl
tqdm
